fix: numbering per heading element
css counters are scoped to the element that first declares them.
This change declares the counters at the content root, so that heading elements
share them regardless of nesting. This also means that we have to clear out all
the lower order counters for every header as they are shared.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>