feat: bring back old ToC behaviour
This is a loving homage to the original ToC. It brings back all the old behaviours and great perf. TocBot perf is terrible on a large doc, so it is removed, in favour of the old ways.

Fixes up resource loading, notably defering katex till well after the page is loaded, to speed up ToC build and reduce the time to get to a useable document.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>