wip: build toc from data.
...parse data from dom. build toc from data.

an experiment in not building the toc client side. instead we build the site without a toc, parse the dom, build the data model, write the toc model to json in the hugo data dir, which then lets us build the toc from data. gross, but it means the toc isnt built in the client, which is worth it, as it explodes on mobile.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>