fix(content): chainsync section update (#1123)
This PR is updating the ChainSync section to match more closely the latest developments. Although nothing was terribly wrong in the previous version of the section, the text looked more like a rather theoretical specification. Furthermore, lots has changed in the way pubsub (i.e., GossipSub) works, which made the chainsync text look out of date.

We will try to keep a version of the previous text and link to it for those that will want to refer back to it.

- Gossip spec references