fix(content): update FIL Token Section (#1155)
This PR applies restructuring to the FIL Token section and adds description to some of the subsections. In particular, the FIL Wallet section is temporarily removed and instead we add the Decay Model and the Token Allocation.

The PR will also update the MultiSig Actor and the pointers to the related repositories.